pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "pysemdist"
version = "0.2.0"
description = "Semantic extraction and API service"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="Your Team"}]
dependencies = [
        "scikit-learn>=1.3.0",
        "sentence-transformers>=2.7.0",
        "hdbscan>=0.8.33",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pandas>=2.0.0",
    "boto3>=1.28.0",
    "psycopg2-binary>=2.9.9"
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.6.0", "mypy>=1.8.0", "httpx>=0.27.0"]

[project.scripts]
pysemdist-api = "pysemdist.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pysemdist"]

[tool.ruff]
line-length = 100
target-version = "py310"
